+++ app/Plugin/Helper.php
@@ -820,21 +820,21 @@ class Helper {
 	 *
 	 * @return array<int,object>
 	 */
-    public static function get_blogs( int $exclude_blog_id = 0 ): array {
+	public static function get_blogs( int $exclude_blog_id = 0 ): array {
 		// query for the list of blogs.
 		global $wpdb;
-	    return $wpdb->get_results(
-		    $wpdb->prepare(
-			    "
+		return $wpdb->get_results(
+			$wpdb->prepare(
+				'
 	            SELECT blog_id
-	            FROM " . $wpdb->blogs . "
-	            WHERE site_id = " . $wpdb->siteid . "
+	            FROM ' . $wpdb->blogs . '
+	            WHERE site_id = ' . $wpdb->siteid . "
 	            AND spam = '0'
 	            AND deleted = '0'
 	            AND archived = '0'
 	            AND blog_id != %d",
-			    $exclude_blog_id
-		    )
-	    );
-    }
+				$exclude_blog_id
+			)
+		);
+	}
 }